+++ CHANGELOG.md
@@ -24,6 +24,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix the computer/user stats table so that it displays the correct values
 - Fix Avoid using the ! negation operator warning in New-AbrADDiagram cmdlet
 - Fix DNS infrastructure collection to query all domain controllers @cse-gh
+- Fix the logic of the Exclude.Domains and Include.Domains options in the
+  Invoke-AsBuiltReport.Microsoft.AD cmdlet to ensure that they work as intended
 
 ### :x: Removed